This adds a new patch to Qt/WebKit (qtwebkit-004-add-prepend-API.patch) that adds a new API that facilitates adding history items so that the browser may be pre-loaded with history items on startup.
Then API was added to llmozlib to access this new feature, as well as new API
to clear the History and dump it to a string for debugging.

Modified ubrowser as well to facilitate testing of new API.

See https://jira.lindenlab.com/jira/browse/DEV-33640
and https://bugs.webkit.org/show_bug.cgi?id=26452
for more info.